+++ public/appModalActions.js
@@ -13,12 +13,13 @@ import { setModalOpen } from "./modalUi.js";
  *
  * @param {{uiState: object}} store - Shared app store.
  * @param {object} modalRefs - Modal DOM references.
- * @param {{onBeforeClosePicker?: Function}} [opts] - Optional callbacks.
+ * @param {{onBeforeClosePicker?: Function, onBeforeOpenModal?: Function}}
+ *   [opts] - Optional callbacks.
  * @returns {{updateModalUi: Function, openModalByKey: Function,
  *   closeModalByKey: Function}}
  */
 export function createModalActions(store, modalRefs, opts = {}) {
-  const { onBeforeClosePicker } = opts;
+  const { onBeforeClosePicker, onBeforeOpenModal } = opts;
 
   const allModalKeys = Object.keys(modalRefs);
 
@@ -30,6 +31,7 @@ export function createModalActions(store, modalRefs, opts = {}) {
   }
 
   function openModalByKey(modalKey) {
+    onBeforeOpenModal?.();
     let nextState = { ...store.uiState, modal: { ...store.uiState.modal } };
     for (const key of allModalKeys) nextState = closeModal(nextState, key);
     store.uiState = openModal(nextState, modalKey);